[caravan-split] Committing separate files for the CaravanSplitting_Proxy, The CaravanSplittingSession and the CaravanSplittingPatches
